# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LONGEVITY API",
    "version": "0.1.0",
    "description": "Health-score MVP — backend API"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionId"
      }
    },
    "schemas": {
...
      "Source": {
        "type": "object",
        "required": [
          "id",
          "kind",
          "adapter",
          "enabled",
          "lastSyncAt",
          "sampleCount"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string"
          },
          "adapter": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "connected": {
            "type": "boolean"
          },
          "syncStatus": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "ok",
              "token_expired",
              "error",
              null
            ]
          },
          "syncError": {
            "type": [
              "string",
              "null"
            ]
          },
          "lastSyncAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "sampleCount": {
            "type": "integer"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/sources": {
      "get": {
        "operationId": "getSources",
        "tags": [
          "Sources"
        ],
        "summary": "List data sources",
        "responses": {
          "200": {
            "description": "Source[]",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Source"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      }
    },
    "/samples/summary": {
      "get": {
        "operationId": "getSamplesSummary",
        "tags": [
          "Samples"
        ],
        "summary": "Get synchronized metrics summary and recent samples",
        "responses": {
          "200": {
            "description": "SamplesSummaryResponse"
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      }
    },
    "/sources/{id}": {
      "patch": {
        "operationId": "patchSource",
        "tags": [
          "Sources"
        ],
        "summary": "Enable/disable source",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "enabled"
                ],
                "properties": {
                  "enabled": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Updated"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/sources/{id}/regenerate": {
      "post": {
        "operationId": "regenerateSource",
        "tags": [
          "Sources"
        ],
        "summary": "Re-seed mock source",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "New sample count"
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      }
    },
...
    "/sources/{id}/disconnect": {
      "delete": {
        "operationId": "disconnectSource",
        "tags": [
          "Sources"
        ],
        "summary": "Remove OAuth credentials, keep samples unless deleteData is true",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deleteData",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Disconnected"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/sources/{id}/samples": {
      "delete": {
        "operationId": "deleteSourceSamples",
        "tags": [
          "Sources"
        ],
        "summary": "Delete all samples for a source",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Samples deleted"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  }
}